+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **Custom skill level**
 - **_Prev/Next [Weapon Buttons] Skip Ammoless Weapons_** setting
 - **_Highlight Current/Pending Weapon_** setting
+- **_Show Thing Hitboxes [in Automap]_** setting
 - **_'BABYMODE'_ cheat**, to toggle ITYTD benefits
 - **Explosion shake intensity percent** setting
 - **Setting to play `DSNOWAY` instead of `DSOOF` when failing to use key-locked triggers**